Crawler: Lower pre-fetch to 4 and workers to 10
With high prefetch counts such as >10 this causes some small issues
with ordering but its nothing major. Mainly occurs when the crawler
instances are started and stopped repeatedly in a short span of time.

Worker count reduced from 40 to 10 as from a cold start (empty queue)
all 40 workers would grab a job and start to make requests to the steam
api at the same time. However, since each worker would make anywhere
between 3-8 calls per user this caused a massive buildup of workers
waiting to call the API. This caused the first 10-20 seconds to have
little to no crawling activity which makes it look like the system is
broken for the user. Instead with 10 workers near instant progress is
made and displayed to the user